Button Toggle works and the form is in it's own component
